unpackPlotProps expects a ComponentClass as a third argument
This gives component writers a place to intercept the computed props
passed to their component. The props will also be available to the higher
level connectToPlot component. Critically this allows components to
short circuit their visibility _before_ they are even rendered.

The restriction is that this is a static class method and therefore
must be purely a function of the higher level component props,
react context and the derived plotProps.

DataSelector takes advantage of the static plotProps API which
simplifies unpackPlotProps by removing data src handling that was
particular to the DataSelector. All DataSrc handling is now confined
to components/DataSelector.js

Section and ConnectToPlot adhere to the new unpackPlotProps API