factor out fullTraceIndex handling into utility
We need both traceIndex and fullTraceIndex within each field so that
unpackPlotProps can properly compute the derived plots from
fullData. Before this commit only TraceAccordion new how to generate
fullTraceIndex which meant it was difficult to test or use a data
field widget outside of the TraceAccordion.

This commit adds functionality to unpackPlotProps such that it checks
if the traceIndex is input either by context OR props and if
fullDataIndex is also passed in. If it is not it computes it using the
factored out function. This behaviour still happens in traceAccordion
as it can do the computation once for all of its children (all fields
point to the single trace)