change API for modifying framework
Instead of modifyPlotProps we have a more general supplyPlotProps. If
a supplyPlotProps is present it is run. Most implementations of
supplyPlotProps will run `unpackPlotProps` first and then alter
plotProps as necessary. However, this enables us to now alter `props`
and `context` _before_ running unpackPlotProps. This will allows us to
"fix" the problem of nested layout fields within a trace context. See
https://github.com/plotly/react-plotly.js-editor/issues/58#issuecomment-345492794
for background.